requirements.txt
streamlit>=1.40.0
groq>=0.13.0
google-genai>=1.0.0
python-dotenv>=1.0.0
